// src/lib/schemas/signupSchema.ts
import * as z from "zod";
import { phoneWithCountrySchema } from "./commons";

export const signupSchema = z.object({
    email: z.email().min(1, { message: "Please input a valid email address" }),
    firstName: z.string().min(1, { message: "Please input your first name" }),
    lastName: z.string().min(1, { message: "Please input your last name" }),
    phoneNumber: phoneWithCountrySchema,
    password: z.string().min(8, { message: "Password must be at least 8 characters"})
})